etcdserver: extract togRPCError() to a separated file
It is used from multiple files in v3rpc package.
